Delete TGT children when TGT is destroyed

+++ cas-server-integration-memcached/src/main/java/org/jasig/cas/ticket/registry/MemCacheTicketRegistry.java
@@ -18,13 +18,14 @@
  */
 package org.jasig.cas.ticket.registry;
 
+import java.util.Map;
 import net.spy.memcached.AddrUtil;
 import net.spy.memcached.MemcachedClient;
 import net.spy.memcached.MemcachedClientIF;
 import org.jasig.cas.ticket.ServiceTicket;
 import org.jasig.cas.ticket.Ticket;
 import org.jasig.cas.ticket.TicketGrantingTicket;
-import org.jasig.cas.ticket.registry.encrypt.AbstractCrypticTicketRegistry;
+import org.jasig.cas.authentication.principal.Service;
 import org.springframework.beans.factory.DisposableBean;
 
 import javax.validation.constraints.Min;
@@ -125,9 +126,27 @@ public final class MemCacheTicketRegistry extends AbstractCrypticTicketRegistry 
             logger.error("Failed adding {}", ticket, e);
         }
     }
+
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public boolean deleteTicket(final String ticketIdToDel) {
         final String ticketId = encodeTicketId(ticketIdToDel);
+        if (ticketId == null) {
+            return false;
+        }
+
+        final Ticket ticket = getTicket(ticketId);
+        if (ticket == null) {
+            return false;
+        }
+
+        if (ticket instanceof TicketGrantingTicket) {
+            logger.debug("Removing ticket children [{}] from the registry.", ticket);
+            deleteChildren((TicketGrantingTicket) ticket);
+        }
+
         logger.debug("Deleting ticket {}", ticketId);
         try {
             return this.client.delete(ticketId).get();
@@ -136,6 +155,27 @@ public final class MemCacheTicketRegistry extends AbstractCrypticTicketRegistry 
         }
         return false;
     }
+
+    /**
+     * Delete the TGT's service tickets.
+     *
+     * @param ticket the ticket
+     */
+    private void deleteChildren(final TicketGrantingTicket ticket) {
+        // delete service tickets
+        final Map<String, Service> services = ticket.getServices();
+        if (services != null && !services.isEmpty()) {
+            for (final Map.Entry<String, Service> entry : services.entrySet()) {
+                try {
+                    this.client.delete(entry.getKey());
+                    logger.trace("Scheduled deletion of service ticket [{}]", entry.getKey());
+                } catch (final Exception e) {
+                    logger.error("Failed deleting {}", entry.getKey(), e);
+                }
+            }
+        }
+    }
+
     @Override
     public Ticket getTicket(final String ticketIdToGet) {
         final String ticketId = encodeTicketId(ticketIdToGet);
